prod_requirements.txt
telethon
requests
google-generativeai
